+++ lib/hover.dart
@@ -1,4 +1,3 @@
 library hover;
 
 export 'src/framework/hover_framework.dart';
-export 'src/widgets/hover_widgets.dart';